Restructured utils
