+++ android/app/src/main/java/com/mindful/android/MainActivity.kt
@@ -18,7 +18,6 @@ import androidx.activity.result.contract.ActivityResultContracts
 import com.mindful.android.helpers.AlarmTasksSchedulingHelper.scheduleMidnightResetTask
 import com.mindful.android.helpers.device.NotificationHelper
 import com.mindful.android.helpers.storage.SharedPrefsHelper
-import com.mindful.android.utils.AppConstants
 import io.flutter.embedding.android.FlutterFragmentActivity
 import io.flutter.embedding.engine.FlutterEngine
 import io.flutter.plugin.common.MethodChannel
