+++ android/app/src/main/java/com/mindful/android/services/accessibility/MindfulAccessibilityService.kt
@@ -28,12 +28,12 @@ import com.mindful.android.helpers.device.PermissionsHelper
 import com.mindful.android.helpers.storage.SharedPrefsHelper
 import com.mindful.android.models.WellBeingSettings
 import com.mindful.android.receivers.DeviceAppsChangedReceiver
-import com.mindful.android.utils.AppConstants.FACEBOOK_PACKAGE
-import com.mindful.android.utils.AppConstants.INSTAGRAM_PACKAGE
-import com.mindful.android.utils.AppConstants.REDDIT_PACKAGE
-import com.mindful.android.utils.AppConstants.SETTINGS_PACKAGE
-import com.mindful.android.utils.AppConstants.SNAPCHAT_PACKAGE
-import com.mindful.android.utils.AppConstants.YOUTUBE_PACKAGE
+import com.mindful.android.AppConstants.FACEBOOK_PACKAGE
+import com.mindful.android.AppConstants.INSTAGRAM_PACKAGE
+import com.mindful.android.AppConstants.REDDIT_PACKAGE
+import com.mindful.android.AppConstants.SETTINGS_PACKAGE
+import com.mindful.android.AppConstants.SNAPCHAT_PACKAGE
+import com.mindful.android.AppConstants.YOUTUBE_PACKAGE
 import com.mindful.android.utils.ThreadUtils
 import org.json.JSONObject
 import java.util.concurrent.ExecutorService
