+++ android/app/src/main/java/com/mindful/android/services/tracking/MindfulTrackerService.kt
@@ -9,7 +9,7 @@ import com.mindful.android.R
 import com.mindful.android.generics.ServiceBinder
 import com.mindful.android.helpers.device.NotificationHelper
 import com.mindful.android.helpers.storage.SharedPrefsHelper
-import com.mindful.android.utils.AppConstants
+import com.mindful.android.AppConstants
 
 class MindfulTrackerService : Service() {
     companion object {
