+++ android/app/src/main/java/com/mindful/android/services/vpn/MindfulVpnService.kt
@@ -21,7 +21,7 @@ import com.mindful.android.R
 import com.mindful.android.generics.ServiceBinder
 import com.mindful.android.helpers.device.NotificationHelper
 import com.mindful.android.helpers.storage.SharedPrefsHelper
-import com.mindful.android.utils.AppConstants
+import com.mindful.android.AppConstants
 import java.io.IOException
 import java.net.InetSocketAddress
 import java.net.SocketAddress
