+++ android/app/src/main/java/com/mindful/android/utils/Extensions.kt
@@ -1,24 +1,68 @@
 package com.mindful.android.utils
 
+import android.database.Cursor
 import java.util.Stack
 
 object Extensions {
 
     /**
-     * Looks at the object at the top of this stack without removing it
-     * from the stack.
+     * Safely returns the element at the top of the stack without removing it.
      *
-     * @return  the object at the top of this stack (the last item
-     *          of the {@code Vector} object) or null if empty.
+     * @return The top element of the stack, or `null` if the stack is empty.
      */
     fun <T> Stack<T>.safePeek(): T? = if (isEmpty()) null else peek()
 
     /**
-     * Removes the object at the top of this stack and returns that
-     * object as the value of this function.
+     * Safely removes and returns the element at the top of the stack.
      *
-     * @return  The object at the top of this stack (the last item
-     *          of the {@code Vector} object) or null if empty.
+     * @return The top element of the stack, or `null` if the stack is empty.
      */
     fun <T> Stack<T>.safePop(): T? = if (isEmpty()) null else pop()
+
+    /**
+     * Retrieves the integer value for the specified column or returns a default value if
+     * the column is missing or the value is invalid.
+     *
+     * @param column The name of the column to retrieve.
+     * @param default The default value to return if the column is not found or fails.
+     * @return The integer value of the column or the default.
+     */
+    fun Cursor.getIntOrDefault(column: String, default: Int = 0): Int {
+        return runCatching { getInt(getColumnIndexOrThrow(column)) }.getOrElse { default }
+    }
+
+    /**
+     * Retrieves the long value for the specified column or returns a default value if
+     * the column is missing or the value is invalid.
+     *
+     * @param column The name of the column to retrieve.
+     * @param default The default value to return if the column is not found or fails.
+     * @return The long value of the column or the default.
+     */
+    fun Cursor.getLongOrDefault(column: String, default: Long = 0L): Long {
+        return runCatching { getLong(getColumnIndexOrThrow(column)) }.getOrElse { default }
+    }
+
+    /**
+     * Retrieves the boolean value for the specified column or returns a default value if
+     * the column is missing or the value is invalid. The column is expected to be stored as an integer (0 or 1).
+     *
+     * @param column The name of the column to retrieve.
+     * @param default The default value to return if the column is not found or fails.
+     * @return The boolean value of the column or the default.
+     */
+    fun Cursor.getBooleanOrDefault(column: String, default: Boolean = false): Boolean {
+        return getIntOrDefault(column, if (default) 1 else 0) == 1
+    }
+
+    /**
+     * Retrieves the string value for the specified column or returns `null` if
+     * the column is missing or an error occurs.
+     *
+     * @param column The name of the column to retrieve.
+     * @return The string value of the column or `null`.
+     */
+    fun Cursor.getStringOrNull(column: String): String? {
+        return runCatching { getString(getColumnIndexOrThrow(column)) }.getOrNull()
+    }
 }
