+++ android/app/src/main/java/com/mindful/android/utils/Utils.kt
@@ -37,49 +37,6 @@ import kotlin.math.abs
 object Utils {
     private const val TAG = "Mindful.Utils"
 
-
-    /**
-     * Creates a intent for Mindful with the provided URI.
-     *
-     * @param context          The application context.
-     * @param uriString          The string representation of URI.
-     * @return The pending intent.
-     */
-    fun getIntentForMindfulUri(
-        context: Context,
-        uriString: String = "com.mindful.android://open/home",
-    ): Intent {
-        val appIntent = Intent(
-            Intent.ACTION_VIEW,
-            Uri.parse(uriString)
-        ).apply {
-            `package` = context.packageName
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        }
-
-        return appIntent
-    }
-
-    /**
-     * Creates a pending intent for Mindful with the provided URI.
-     *
-     * @param context          The application context.
-     * @param uriString          The string representation of URI.
-     * @return The pending intent.
-     */
-    fun getPendingIntentForMindfulUri(
-        context: Context,
-        uriString: String = "com.mindful.android://open/home",
-    ): PendingIntent {
-
-        return PendingIntent.getActivity(
-            context.applicationContext,
-            0,
-            getIntentForMindfulUri(context, uriString),
-            PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-        )
-    }
-
     /**
      * Checks if a service with the given class name is currently running.
      *
@@ -96,231 +53,6 @@ object Utils {
         }
 
         return false
-    }
-
-    /**
-     * Get the current app version.
-     *
-     * @param context The context from which the method is called
-     * @return The app version as a string
-     */
-    fun getAppVersion(context: Context): String {
-        try {
-            val packageManager = context.packageManager
-            val packageName = context.packageName
-            val packageInfo = packageManager.getPackageInfo(packageName, 0)
-
-            return if (packageName.contains(".debug")
-            ) "v" + packageInfo.versionName + "-debug+" + packageInfo.versionCode
-            else "v" + packageInfo.versionName + "+" + packageInfo.versionCode
-        } catch (e: Exception) {
-            return "unknown"
-        }
-    }
-
-    /**
-     * Resolve the device information and  returns it
-     *
-     * @return Map containing Manufacturer, Model, Android Version, SDK Version and Mindful version.
-     */
-    fun getDeviceInfoMap(context: Context): Map<String, Any> {
-        val infoMap = HashMap<String, Any>()
-        infoMap["manufacturer"] = Build.MANUFACTURER
-        infoMap["model"] = Build.MODEL
-        infoMap["androidVersion"] = Build.VERSION.RELEASE
-        infoMap["sdkVersion"] = Build.VERSION.SDK_INT
-        infoMap["mindfulVersion"] = getAppVersion(context)
-        return infoMap
-    }
-
-
-    /**
-     * Encodes a Drawable (app icon) to a Base64 string.
-     *
-     * @param iconData The Drawable to encode.
-     * @return The Base64 encoded string representing the app icon.
-     */
-    private fun encodeToBase64(iconData: Drawable): String {
-        val byteArrayOS = ByteArrayOutputStream()
-        val image = Bitmap.createBitmap(
-            iconData.intrinsicWidth,
-            iconData.intrinsicHeight,
-            Bitmap.Config.ARGB_8888
-        )
-
-        val canvas = Canvas(image)
-        iconData.setBounds(0, 0, canvas.width, canvas.height)
-        iconData.draw(canvas)
-        image.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOS)
-        return Base64.encodeToString(byteArrayOS.toByteArray(), Base64.NO_WRAP)
-    }
-
-    /**
-     * Encodes and returns the Base64 representation of an app icon.
-     *
-     * @param icon The app icon as a Drawable.
-     * @return The Base64 encoded string of the app icon.
-     */
-    fun getEncodedAppIcon(icon: Drawable): String {
-        var appIcon = ""
-        try {
-            appIcon = encodeToBase64(icon)
-        } catch (e: Exception) {
-            Log.e(TAG, "getEncodedAppIcon: Cannot parse app icon from memory", e)
-        }
-        return appIcon
-    }
-
-    /**
-     * Converts the drawable to bitmap.
-     *
-     * @param drawable The Drawable which will be converted.
-     * @return Bitmap The converted bitmap.
-     */
-    fun drawableToBitmap(drawable: Drawable): Bitmap {
-        if (drawable is BitmapDrawable) {
-            return drawable.bitmap
-        }
-
-        val bitmap = Bitmap.createBitmap(
-            drawable.intrinsicWidth,
-            drawable.intrinsicHeight,
-            Bitmap.Config.ARGB_8888
-        )
-        val canvas = Canvas(bitmap)
-        drawable.setBounds(0, 0, canvas.width, canvas.height)
-        drawable.draw(canvas)
-
-        return bitmap
-    }
-
-    /**
-     * Formats the total screen usage time into a human-readable string.
-     * Example: 02h 43m 09s
-     *
-     * @param totalSeconds The total screen usage time in seconds.
-     * @return A string representing the formatted screen usage time.
-     */
-    @Contract(pure = true)
-    fun secondsToTimeStr(totalSeconds: Int): String {
-        val leftHours = (totalSeconds / 3600)
-        val leftMinutes = ((totalSeconds % 3600) / 60)
-        val leftSeconds = (totalSeconds % 60)
-
-        val timeStr = StringBuilder()
-
-        if (leftHours > 0) {
-            timeStr.append(String.format(Locale.ENGLISH, "%02d", leftHours)).append("h ")
-        }
-
-        timeStr.append(String.format(Locale.ENGLISH, "%02d", leftMinutes)).append("m ")
-        timeStr.append(String.format(Locale.ENGLISH, "%02d", leftSeconds)).append("s")
-
-        return timeStr.toString()
-    }
-
-    /**
-     * Returns the current day of the week as a zero-based index, where Monday is 0 and Sunday is 6.
-     *
-     * The `Calendar.DAY_OF_WEEK` values range from `1` (Sunday) to `7` (Saturday).
-     * This method shifts the values so that:
-     * - Monday → `0`
-     * - Tuesday → `1`
-     * - Wednesday → `2`
-     * - Thursday → `3`
-     * - Friday → `4`
-     * - Saturday → `5`
-     * - Sunday → `6`
-     *
-     * @return An integer representing the day of the week with Monday as index `0` and Sunday as index `6`.
-     */
-    @Contract(pure = true)
-    fun zeroIndexedDayOfWeek(): Int {
-        val javaDayOfWeek = Calendar.getInstance()[Calendar.DAY_OF_WEEK]
-        return (javaDayOfWeek + 5) % 7
-    }
-
-
-    /**
-     * Create a calender instance from the Time Of Day total minutes.
-     * If totalMinutes is negative then the calender is adjusted to previous day
-     *
-     * @param totalMinutes The total minutes from Time Of Day dart object.
-     * @return Calender representing the time of day for today.
-     */
-    @Contract(pure = true)
-    fun todToTodayCal(totalMinutes: Int): Calendar {
-        val cal = Calendar.getInstance()
-        cal[Calendar.HOUR_OF_DAY] = 0
-        cal.add(Calendar.HOUR_OF_DAY, totalMinutes / 60)
-        cal[Calendar.MINUTE] = 0
-        cal.add(Calendar.MINUTE, totalMinutes % 60)
-        cal[Calendar.SECOND] = 0
-        cal[Calendar.MILLISECOND] = 0
-        return cal
-    }
-
-    /**
-     * Calculated the difference between time now and future tod minutes.
-     *
-     * @param futureTotalMinutes The total minutes from Time Of Day dart object.
-     * @return The difference in MS. If the difference is negative then return 0
-     */
-    fun todDifferenceFromNow(futureTotalMinutes: Int): Long {
-        val diff = todToTodayCal(futureTotalMinutes).timeInMillis - System.currentTimeMillis()
-        return if (diff > 0) diff else 0
-    }
-
-    /**
-     * Checks if current system time is outside the star and end Time Of Day for today
-     *
-     * @param startTod The start TOD in minutes
-     * @param endTod   The end TOD in minutes
-     * @return Boolean indicating if now is outside startTod and endTod
-     */
-    fun isTimeOutsideTODs(startTod: Int, endTod: Int): Boolean {
-        val now = System.currentTimeMillis()
-        return !(todToTodayCal(startTod).timeInMillis < now && now < todToTodayCal(endTod).timeInMillis)
-    }
-
-
-    /**
-     * Formats the total screen usage time into a human-readable string.
-     * Example: 12h 5m
-     *
-     * @param totalMinutes The total screen usage time in minutes.
-     * @return A string representing the formatted screen usage time.
-     */
-    @Contract(pure = true)
-    fun minutesToTimeStr(totalMinutes: Int): String {
-        val totalMinutesAbs = abs(totalMinutes)
-
-        return if (totalMinutesAbs > 60
-        ) if (totalMinutesAbs % 60 == 0
-        ) (totalMinutesAbs / 60).toString() + "h"
-        else (totalMinutesAbs / 60).toString() + "h " + totalMinutesAbs % 60 + "m"
-        else totalMinutesAbs.toString() + "m"
-    }
-
-
-    /**
-     * Formats the total data usage into a human-readable string.
-     * Example: 12.35 GB
-     *
-     * @param totalMBs The total data usage in megabytes (MB).
-     * @return A string representing the formatted data usage.
-     */
-    @Contract(pure = true)
-    fun formatDataMBs(totalMBs: Int): String {
-        val totalMBsAbs = abs(totalMBs)
-
-        if (totalMBsAbs >= 1024) {
-            val GBs = totalMBsAbs / 1024f
-            val formattedGBs = String.format(Locale.getDefault(), "%.2f", GBs)
-            return "$formattedGBs GB"
-        } else {
-            return "$totalMBsAbs MB"
-        }
     }
